refactor(site): cut over to the standalone GraphQL api (HTTP consumer)
Site no longer runs PostGraphile in-process. Server-side GraphQL (RSC, SSR,
generateStaticParams, generateMetadata) now fetches the packages/api service
over HTTP, same as the browser — one less schema build in the Next process.

Site:
- apollo-rsc.ts: in-process grafast link → HttpLink at the api URL, with the
  server-side x-vercel-protection-bypass fetch wrapper (preview SSO) reinstated;
  cache stays plain (no typePolicies). Throws on Vercel if the api URL is unset.
- apollo-make-client.ts (browser): fallback → the local api URL.
- delete server-graphql.ts, postgraphileMiddleware.ts, runMiddleware.ts, and the
  pages/api graphql/graphiql/stream/ruru-static routes (all in-process surface).
- delete the now-duplicated graphile.config.ts + F1dbSmartTags + WikipediaBioPlugin
  (live in packages/api as of the scaffold commit).
- next.config.mjs: drop the postgraphile serverExternalPackages + webpack
  externals block. Remove postgraphile + @graphile/simplify-inflection from deps
  (pg stays for ingest until it moves in the next chunk).

api:
- server.ts builds + validates the schema at boot (ports the M3 fail-fast guard
  that died with server-graphql.ts): asserts seasons/drivers/teams/circuits,
  exit(1) with the POSTGRES_SCHEMA diagnostic on miss. Boot-time build also
  retires the M2 rejected-promise concern and warms the first query.

Verified vs the live api on :4000: site tsc clean, jest 26/26, next build green
(all SSG/PPR pages fetch over HTTP, graphql routes gone), and the api refuses to
listen on a wrong POSTGRES_SCHEMA. Env-file URLs (.env.production still relative)
are Phase C.
