+++ packages/site/package.json
@@ -57,7 +57,6 @@
 		"@fortawesome/free-brands-svg-icons": "^7.2.0",
 		"@fortawesome/free-solid-svg-icons": "^7.2.0",
 		"@fortawesome/react-fontawesome": "^3.3.1",
-		"@graphile/simplify-inflection": "^8.0.0",
 		"@graphql-typed-document-node/core": "^3.2.0",
 		"@gtibrett/mui-additions": "0.7.0-alpha.1718858125",
 		"@mui/material": "^9.0.1",
@@ -87,7 +86,6 @@
 		"i18n-nationality": "^1.4.0",
 		"next": "^16.2.6",
 		"pg": "^8.21.0",
-		"postgraphile": "^5.0.3",
 		"react": "^19.2.6",
 		"react-dom": "^19.2.6",
 		"rxjs": "^7.8.2",
